# swagger/api.yaml
openapi: 3.0.0
info:
  description: "This is the static API documentation for the project called medical-metadata which deals as a proof of concept for the master thesis of Bence Slajchó @Alpen-Adria-University-Klagenfurt. Supervisors: O.Univ.-Prof. Dipl.-Ing. Dr. Johann Eder; Dr.phil. Volodymyr Shekhovtsov"
  version: "1.3.0"
  title: "Medical Metadata"
  contact:
    email: "slajchobence@gmail.com"
...
servers:
  - url: localhost:3001
    description: development server
paths:
...
  /api/loinc-value-ranges:
    get:
      tags:
      - search
      summary: "returns information of available value ranges for the loinc code"
      description: "For Qt scaled LOINC code the bounds are returned"
      parameters:
        - in: query
          name: loincCode
          required: true
          schema:
            type: string
            example: 2161-8
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseValueRangeQt"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # LOINC values ranges for nom
  /api/column-values-nominal:
    get:
      tags:
      - search
      summary: "returns information of available nominal values for the loinc code"
      description: "For Nom scaled LOINC code the avalable values are returned"
      parameters:
        - in: query
          name: loincCode
          required: true
          schema:
            type: string
            example: 2161-8
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseValuesNom"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # LOINC values ranges for ord
  /api/column-values-ordinal:
    get:
      tags:
      - search
      summary: "returns information of available ordnal values for the loinc code"
      description: "For Ord scaled LOINC code the avalable values are returned"
      parameters:
        - in: query
          name: loincCode
          required: true
          schema:
            type: string
            example: 5794-3
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseValuesOrd"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # Meta data delivering
  /api/meta-data:
    get:
      tags:
      - search
      summary: "Returns content based meta information of the loinc concept in the meta-level"
      description: "This route utilizes the content based descriptors and delivers information about a property of a collection. Depending on the scaling additional parameters can be required"
      parameters:
        - in: query
          name: loinc_code
          required: true
          schema:
            type: string
            example: "2161-8"
        - in: query
          name: scaling
          required: true
          schema:
            type: string
            example: "QN"
        - in: query
          name: lower_bound
          required: false
          schema:
            type: string
            example: "0"
        - in: query
          name: upper_bound
          required: false
          schema:
            type: string
            example: "28.2"
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ApiResponseMetadataContentDescriptorQt"
                  - $ref: "#/components/schemas/ApiResponseMetadataContentDescriptorNom"
                  - $ref: "#/components/schemas/ApiResponseMetadataContentDescriptorOrd"
                  - $ref: "#/components/schemas/ApiResponseMetadataContentDescriptorDoc"
                  - $ref: "#/components/schemas/ApiResponseMetadataContentDescriptorNar"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
  # metadata creation and deletion
...
  /api/composite-content-descriptor:
    get:
      tags:
      - "composite descriptor"
      summary: "list of existing composite content descriptors"
      description: "composite content descriptors are such descriptors which can be defined by admins in the system. Such descriptors contain possibly multiple but at least one LOINC code. 
        Furthermore they also contain an operator which can be 'ANY' or 'ALL'. 
        ANY would apply at the search in the following way: if the composite key contains LOINC code 1 and LOINC code 2, then any collection would be returned by the query which contain either LOINC code 1 OR LOINC code 2 (maybe both). Example: 'Find all collections which contain any primary diagnosis'
        ALL would apply at the search in the following way: if the composite key contains LOINC code 1 and LOINC code 2, then only those collections will be returned which contain both LOINC code 1 AND LOINC code 2. Example: 'Find all collections which contain BMI related information'" 
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCompositeContentDescriptorsGet"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
    delete:
      tags:
      - "composite descriptor"
      summary: "deleting existing composite content descriptors by their id-s"
      description: "As a pre-requisite the ID of the composite collection descriptor must be emitted with the get call of this route. Then the the id in a form of an array
        must be sent to the delete route and the those composite content descriptors will be deleted "
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idList:
                  type: array
                  items: 
                    type: integer
                    example: [1, 2]
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCompositeContentDescriptorsDelete"
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"  
    post:
      tags:
      - "composite descriptor"
      summary: "creating new composite content descriptors"
      description: "To create composite content descriptors this route must be called with the required parameters.
        Composite content descriptors are such descriptors which can be defined by admins in the system. Such descriptors contain possibly multiple but at least one LOINC code. 
        Furthermore they also contains an operator which can be 'ANY' or 'ALL'. 
        ANY would apply at the search in the following way: if the composite key contains LOINC code 1 and LOINC code 2, then any collection would be returned by the query which contain either LOINC code 1 OR LOINC code 2 (maybe both). Example: 'Find all collections which contain any primary diagnosis'
        ALL would apply at the search in the following way: if the composite key contains LOINC code 1 and LOINC code 2, then only those collections will be returned which contain both LOINC code 1 AND LOINC code 2. Example: 'Find all collections which contain BMI related information'"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                compositeContentDescriptorName:
                  type: string
                  example: "Test descriptor"
                compositeContentDescriptorDescription:
                  type: string
                  example: "Test descriptor description which will be displayed as help test at search"
                operator:
                  type: string
                  example: "ANY"
                participatingLoincCodes:
                  type: array
                  items: 
                    type: object
                    properties:
                      loinc_code:
                        type: string
                  example: [
                    { loinc_code: "53120-2"},
                    { loinc_code: "33352-6"},
                    { loinc_code: "53561-7"},
                    { loinc_code: "48995-5"},
                    { loinc_code: "2161-8"},
                  ]
      responses:
        "200":
          description: "successful request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCompositeContentDescriptorsPost"
        "409":
          description: "conflict - composite keys exists already"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseErrorCompositeKeyExists"  
        "500":
          description: "server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
...
components:
  schemas:
...
    ApiResponseMetadataContentDescriptorQt:
      type: "array"
      items:
        type: "object"
        properties:
          meta:
            $ref: "#/components/schemas/aggregateMetaInfoQt"
    ApiResponseMetadataContentDescriptorNom:
      type: "array"
      items:
        type: "object"
        properties:
          meta:
            $ref: "#/components/schemas/aggregateMetaInfoNom"
    ApiResponseMetadataContentDescriptorOrd:
      type: "array"
      items:
        type: "object"
        properties:
          meta:
            $ref: "#/components/schemas/aggregateMetaInfoOrd"
    ApiResponseMetadataContentDescriptorDoc:
      type: "array"
      items:
        type: "object"
        properties:
          meta:
            $ref: "#/components/schemas/aggregateMetaInfoDoc"
    ApiResponseMetadataContentDescriptorNar:
      type: "array"
      items:
        type: "object"
        properties:
          meta:
            $ref: "#/components/schemas/aggregateMetaInfoNar"
    aggregateMetaInfoQt:
      type: array
      items: 
        type: object
        properties:
          aggregate_value: 
            type: number
            example: 0
          occurance_count: 
            type: number
            example: 1
          source_database:
            type: string
            example: "raw_medical_data_1"
          value_unit: 
            type: string
            example: "mg/dL"
          min_value: 
            type: number
            example: 0
          max_value:
            type: number
            example: 2
          mean_value: 
            type: number
            example: 1
          median_value: 
            type: number
            example: 1.2
          min_updaed_at: 
            type: string
            format: date-time
            example: Sun Oct 06 2019 01:07:03 GMT+0000 (Coordinated Universal Time)
          max_updaed_at: 
            type: string
            format: date-time
            example: Thu Dec 27 2018 13:37:52 GMT+0000 (Coordinated Universal Time)
          mean_updaed_at: 
            type: string
            format: date-time
            example: Fri May 17 2019 19:22:27 GMT+0000 (Coordinated Universal Time)
          median_updaed_at: 
            type: string
            format: date-time
            example: Thu Dec 27 2018 13:37:52 GMT+0000 (Coordinated Universal Time)
    aggregateMetaInfoNom:
      type: array
      items: 
        type: object
        properties:
          aggregate_value: 
            type: number
            example: 0
          occurance_count: 
            type: number
            example: 1
          source_database:
            type: string
            example: "raw_medical_data_1"
          min_updaed_at: 
            type: string
            format: date-time
            example: Sun Oct 06 2019 01:07:03 GMT+0000 (Coordinated Universal Time)
          max_updaed_at: 
            type: string
            format: date-time
            example: Thu Dec 27 2018 13:37:52 GMT+0000 (Coordinated Universal Time)
          mean_updaed_at: 
            type: string
            format: date-time
            example: Fri May 17 2019 19:22:27 GMT+0000 (Coordinated Universal Time)
          median_updaed_at: 
            type: string
            format: date-time
            example: Thu Dec 27 2018 13:37:52 GMT+0000 (Coordinated Universal Time)
    aggregateMetaInfoOrd:
      type: array
      items: 
        type: object
        properties:
          aggregate_value: 
            type: number
            example: 0
          occurance_count: 
            type: number
            example: 1
          source_database:
            type: string
            example: "raw_medical_data_1"
          min_updaed_at: 
            type: string
            format: date-time
            example: Sun Oct 06 2019 01:07:03 GMT+0000 (Coordinated Universal Time)
          max_updaed_at: 
            type: string
            format: date-time
            example: Thu Dec 27 2018 13:37:52 GMT+0000 (Coordinated Universal Time)
          mean_updaed_at: 
            type: string
            format: date-time
            example: Fri May 17 2019 19:22:27 GMT+0000 (Coordinated Universal Time)
          median_updaed_at: 
            type: string
            format: date-time
            example: Thu Dec 27 2018 13:37:52 GMT+0000 (Coordinated Universal Time)
    aggregateMetaInfoDoc:
      type: array
      items: 
        type: object
        properties:
          aggregate_value: 
            type: number
            example: 0
          occurance_count: 
            type: number
            example: 1
          source_database:
            type: string
            example: "raw_medical_data_1"
          min_updaed_at: 
            type: string
            format: date-time
            example: Sun Oct 06 2019 01:07:03 GMT+0000 (Coordinated Universal Time)
          max_updaed_at: 
            type: string
            format: date-time
            example: Thu Dec 27 2018 13:37:52 GMT+0000 (Coordinated Universal Time)
          mean_updaed_at: 
            type: string
            format: date-time
            example: Fri May 17 2019 19:22:27 GMT+0000 (Coordinated Universal Time)
          median_updaed_at: 
            type: string
            format: date-time
            example: Thu Dec 27 2018 13:37:52 GMT+0000 (Coordinated Universal Time)
    aggregateMetaInfoNar:
      type: array
      items: 
        type: object
        properties:
          aggregate_value: 
            type: number
            example: 0
          occurance_count: 
            type: number
            example: 1
          source_database:
            type: string
            example: "raw_medical_data_1"
          min_updaed_at: 
            type: string
            format: date-time
            example: Sun Oct 06 2019 01:07:03 GMT+0000 (Coordinated Universal Time)
          max_updaed_at: 
            type: string
            format: date-time
            example: Thu Dec 27 2018 13:37:52 GMT+0000 (Coordinated Universal Time)
          mean_updaed_at: 
            type: string
            format: date-time
            example: Fri May 17 2019 19:22:27 GMT+0000 (Coordinated Universal Time)
          median_updaed_at: 
            type: string
            format: date-time
            example: Thu Dec 27 2018 13:37:52 GMT+0000 (Coordinated Universal Time)
...
    ApiResponseError:
      type: "object"
      properties:
        status:
          type: "integer"
          format: "123"
          enum: [500]
        message:
          type: "string"
    ApiResponseErrorCompositeKeyExists:
      type: "object"
      properties:
        status:
          type: "integer"
          format: "123"
          enum: [409]
        error:
          type: "string"
...
    ApiResponseValueRangeQt:
      type: object
      properties:
        min_value:
          type: number
          example: 0
        max_value:
          type: number
          example: 100
    ApiResponseValuesNom:
      type: object
      properties:
        distinct_values:
          type: string
          example: 'amber'
    ApiResponseValuesOrd:
      type: object
      properties:
        sequence_number:
          type: number
          example: 1
        value_set_value:
          type: string
          example: '1'
...
    ApiResponseCompositeContentDescriptorsGet:
      type: "array"
      items:
        type: object
        properties:
          code_operator:
            type: string
            example: "ANY"
            enum: ["ANY", "ALL"]
          description:
            type: string
            example: "The test composite decriptor returns collections which satisfy the decriptor"
          descriptor_name:
            type: string
            example: "Test composite content descriptor"
          id:
            type: number
            example: 1
          participating_codes:
            type: array
            items:
              type: string
              example: ["5794-3"]
    ApiResponseCompositeContentDescriptorsDelete:
      type: object
      properties:
        message:
          type: string
          example: "deleted"
          enum: ["deleted"]
    ApiResponseCompositeContentDescriptorsPost:
      type: object
      properties:
        insertedRowCount:
          type: integer
          example: 1
...